refactor(src, public): refactored classes to move data generation into data module, filters have been replaced with single filter form
