+++ public/index.html
@@ -26,9 +26,6 @@
       <nav class="trip-controls__menus view-switch">
         <a href="#table" class="view-switch__item view-switch__item--active">Table</a>
         <a href="#stats" class="view-switch__item">Stats</a>
-
-        <form class="trip-filter">
-        </form>
       </nav>
 
       <button class="trip-controls__new-event new-event">+ New Event</button>
