+++ src/data.js
@@ -1,4 +1,4 @@
-import {getRandomIntegerFromInterval, replaceText, FILTERS_NAMES, Units} from "./constants";
+import {getRandomIntegerFromInterval, replaceText, FILTERS_NAMES, OFFERS, TYPES, Units} from "./constants";
 
 export let dataArray = [];
 
@@ -26,14 +26,22 @@ const citiesAmount = 1;
 const weeksAmount = 12;
 const minPrice = 20;
 const maxPrice = 2500;
+const maxOfferPrice = 100;
 
 const sortAndRandomlySliceArray = (array, maxAmount) => {
   return array.sort(() =>(Math.random() - 0.5)).slice(0, maxAmount);
 };
 
-const getRandomOffers = (offersSet, maxAmount) => {
-  const offers = [...offersSet];
-  return new Set(sortAndRandomlySliceArray(offers, maxAmount));
+const getOffers = () => {
+  let offers = {};
+  OFFERS.forEach((offer) => {
+    offers[offer] = getRandomIntegerFromInterval(minPrice, maxOfferPrice);
+  });
+  return offers;
+};
+
+const getRandomOffers = (maxAmount) => {
+  return new Set(sortAndRandomlySliceArray(OFFERS, maxAmount));
 };
 
 const createRandomText = (sentencesArray, sentencesAmount) => {
@@ -42,6 +50,10 @@ const createRandomText = (sentencesArray, sentencesAmount) => {
 
 const getNearestWeekendRandomDate = () => {
   return Date.now() + Math.floor(getRandomIntegerFromInterval(0, Units.daysInWeek * weeksAmount + 1)) * Units.hoursInDay * Units.minutesInHour * Units.secondsInMinute * Units.millisecondsInSecond;
+};
+
+const getRandomType = (typesObj) => {
+  return Object.keys(typesObj)[getRandomIntegerFromInterval(0, Object.keys(typesObj).length - 1)];
 };
 
 const getTimeFromDate = (date) => {
@@ -54,26 +66,11 @@ const getTimeFromDate = (date) => {
 
 const getCardData = () => {
   return {
-    type: {
-      'Taxi': `🚕`,
-      'Bus': `🚌`,
-      'Train': `🚂`,
-      'Ship': `🛳️`,
-      'Transport': `🚊`,
-      'Drive': `🚗`,
-      'Flight': `✈️`,
-      'Check-in': `🏨`,
-      'Sightseeing': `🏛️`,
-      'Restaurant': `🍴`,
-    },
+    type: getRandomType(TYPES),
     city: [`Amsterdam`, `Geneve`, `Chamonix`, `Maskat`, `Rio de Janeiro`, `Kioto`][getRandomIntegerFromInterval(0, citiesAmount)],
     picture: `http://picsum.photos/300/150?r=${Math.random()}`,
-    offers: getRandomOffers(new Set([
-      `Add luggage`,
-      `Switch to comfort class`,
-      `Add meal`,
-      `Choose seats`
-    ]), 2),
+    offers: getOffers(),
+    chosenOffers: getRandomOffers(2),
     text: createRandomText([`Lorem ipsum dolor sit amet, consectetur adipiscing elit.`, `Cras aliquet varius magna, non porta ligula feugiat eget.`, `Fusce tristique felis at fermentum pharetra.`, `Aliquam id orci ut lectus varius viverra.`, `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`, `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`, `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`, `Sed sed nisi sed augue convallis suscipit in sed felis.`, `Aliquam erat volutpat.`, `Nunc fermentum tortor ac porta dapibus.`, `In rutrum ac purus sit amet tempus.`], getRandomIntegerFromInterval(minTextLength, maxTextLength)),
     time: getTimeFromDate(getNearestWeekendRandomDate()),
     price: getRandomIntegerFromInterval(minPrice, maxPrice)
