+++ src/main.js
@@ -1,29 +1,54 @@
-import {getRandomIntegerFromInterval} from "./constants";
-import {FilterInput, FilterLabel} from "./filters";
-import {renderPoints} from "./points";
+import {getRandomIntegerFromInterval, Units} from "./constants";
+import {Filter} from "./filter";
+import {Point} from "./points";
 import {dataArray, filtersData, randomizeData} from "./data";
+import {PointEdit} from "./pointEdit";
 
 let fragment = document.createDocumentFragment();
 
-const filtersRoot = document.querySelector(`.trip-filter`);
+const filterRoot = document.querySelector(`.trip-controls__menus`);
+const pointsRoot = document.querySelector(`.trip-day__items`);
 
-const init = () => {
+const renderPoints = (data) => {
+  pointsRoot.innerHTML = ``;
+  fragment = document.createDocumentFragment();
   let i = 0;
-  while (i < filtersData.length) {
-    let filterInput = new FilterInput(filtersData[i]);
-    let filterLabel = new FilterLabel(filtersData[i]);
-    filterLabel.clickHandler = () => {
-      renderPoints(randomizeData(getRandomIntegerFromInterval(7, 20)));
+  while (i < data.length) {
+    let point = new Point(data[i]);
+    let pointEdit = new PointEdit(data[i]);
+    point.editHandler = () => {
+      pointEdit.render();
+      pointsRoot.replaceChild(pointEdit.element, point.element);
+      point.unrender();
     };
-    fragment.appendChild(filterInput.render());
-    fragment.appendChild(filterLabel.render());
+
+    pointEdit.submitHandler = () => {
+      point.render();
+      pointsRoot.replaceChild(point.element, pointEdit.element);
+      pointEdit.unrender();
+    };
+
+    pointEdit.resetHandler = () => {
+      pointsRoot.removeChild(pointEdit.element);
+      pointEdit.unrender();
+    };
+
+    fragment.appendChild(point.render());
     i++;
   }
+
+  pointsRoot.appendChild(fragment);
+};
+
+const init = () => {
+  let filters = new Filter(filtersData);
+  filters.clickHandler = () => {
+    renderPoints(randomizeData(getRandomIntegerFromInterval(Units.startUnit, 7)));
+  };
+  fragment.appendChild(filters.render());
+  filterRoot.appendChild(fragment);
 
   renderPoints(dataArray);
 };
 
 init();
-
-
-filtersRoot.appendChild(fragment);
